Fix review comments: - fix naming; -  move event listener for onMovieClieck at filmComponent level;
